Define input* and output* methods by data type in class DataUtils in package edu.utdallas.davisbase.storage
While converting a value of a known type to a byte array is possible, the reverse is not since the type isn't known just from bytes.  Moreover, the number of bytes required for deserialization of any given type is an implementation detail.  Thus, the decision was made to design output* methods rather than convert*FromBytes methods in order to work around this.  As a minor consequence, the decision was also made to implement analogous input* methods.